api/requirements.txt
flask>=1.1.1
flask-cors>=3.0.8
essentia>=2.1b6.dev184
azure-functions>=1.2.1
azf-wsgi==0.3.0